// src/app/models/empresa.ts
export class Empresa {
    emprId?: Number;
    emprTipoDocumento: String | undefined;
    emprDocumento: String | undefined;
    emprNombreCompleto: String | undefined;
    emprDireccion: String | undefined;
    emprCiudad: String | undefined;
    emprDepartamento: String | undefined;
    emprPais: String | undefined;
    emprTelefono: String | undefined;
    emprLegalTipoDocumento: String | undefined;
    emprLegalDocumento: String | undefined;
    emprLegalNombreCompleto: String | undefined;
    emprLegalDireccion: String | undefined;
    emprLegalCiudad: String | undefined;
    emprLegalDepartamento: String | undefined;
    emprLegalPais: String | undefined;
    emprLegalTelefono: String | undefined;
}